requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.12
django-cloudinary-storage==0.3.0
django-crontab==0.7.1
django-heroku==0.3.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.2
kombu==5.2.4
matplotlib==3.2.2
mysqlclient==2.1.0
nltk==3.7
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.0
regex==2022.3.15
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
Wand==0.6.7
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
